Add event signatures + refactor ambient code
